requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.114.0
fonttools==4.53.1
frozendict==2.4.4
h11==0.14.0
html5lib==1.1
idna==3.8
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
multitasking==0.0.11
numpy==2.1.1
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.1
plotly==5.24.0
pydantic==2.9.0
pydantic_core==2.23.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.4
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
webencodings==0.5.1
yfinance==0.2.43
gunicorn==20.1.0
whitenoise==6.4.0
python-dotenv==1.0.0
